+++ src/Routes/index.js
@@ -15,6 +15,7 @@ const { routerLogin } = require("./loginRoute");
 
 const { routerLoginbackOffice } = require("./routerLoginbackOffice");
 const { routerReservation } = require("./routeReservation");
+const routerMercadoPago = require("./mercadopagoRouter");
 
 const router = Router();
 
@@ -46,5 +47,5 @@ router.use("/loginBackOffice", routerLoginbackOffice);
 router.use("/auth", routerLogin);
 
 router.use("/reservation", routerReservation);
-
+router.use("/pagos", routerMercadoPago);
 module.exports = router;
